requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
py-bcrypt==0.4
python-dotenv==0.17.0
Werkzeug==1.0.1
